Modify parameter type in VolumeInfo.kt

+++ app/src/main/java/com/compose/jreader/data/model/VolumeInfo.kt
@@ -3,7 +3,7 @@ package com.compose.jreader.data.model
 data class VolumeInfo(
     val allowAnonLogging: Boolean?,
     val authors: List<String>?,
-    val averageRating: Int?,
+    val averageRating: Double?,
     val canonicalVolumeLink: String?,
     val categories: List<String>?,
     val contentVersion: String?,
